Refactor: 목록으로 돌아가기 UI를 하나의 컴포넌트로 분리한다
