+++ src/components/apply/index.ts
@@ -5,3 +5,4 @@ export { default as QuestionAndAnswerList } from './QuestionAndAnswerList/Questi
 export { default as BlockingConfirmModalDialog } from './BlockingConfirmModalDialog/BlockingConfirmModalDialog.component';
 export { default as SubmitModalDialog } from './SubmitModalDialog/SubmitModalDialog.component';
 export { default as TempSaveModalDialog } from './TempSaveModalDialog/TempSaveModalDialog.component';
+export { default as BackToListLink } from './BackToListLink/BackToListLink.component';
